[Refactor](Sink) convert to tablet sink to tablet writer
