+++ be/src/exec/data_sink.cpp
@@ -31,7 +31,6 @@
 
 #include "common/config.h"
 #include "vec/sink/async_writer_sink.h"
-#include "vec/sink/group_commit_vtablet_sink.h"
 #include "vec/sink/multi_cast_data_stream_sink.h"
 #include "vec/sink/vdata_stream_sender.h"
 #include "vec/sink/vmemory_scratch_sink.h"
@@ -149,9 +148,10 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
         Status status;
         DCHECK(thrift_sink.__isset.olap_table_sink);
         if (state->query_options().enable_memtable_on_sink_node) {
-            sink->reset(new stream_load::VOlapTableSinkV2(pool, row_desc, output_exprs, &status));
+            sink->reset(new vectorized::VOlapTableSinkV2(pool, row_desc, output_exprs, &status));
         } else {
-            sink->reset(new stream_load::VOlapTableSink(pool, row_desc, output_exprs, &status));
+            sink->reset(
+                    new vectorized::VOlapTableSink(pool, row_desc, output_exprs, false, &status));
         }
         RETURN_IF_ERROR(status);
         break;
@@ -159,8 +159,7 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
     case TDataSinkType::GROUP_COMMIT_OLAP_TABLE_SINK: {
         Status status;
         DCHECK(thrift_sink.__isset.olap_table_sink);
-        sink->reset(
-                new stream_load::GroupCommitVOlapTableSink(pool, row_desc, output_exprs, &status));
+        sink->reset(new vectorized::VOlapTableSink(pool, row_desc, output_exprs, true, &status));
         RETURN_IF_ERROR(status);
         break;
     }
@@ -297,9 +296,10 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
         Status status;
         DCHECK(thrift_sink.__isset.olap_table_sink);
         if (state->query_options().enable_memtable_on_sink_node) {
-            sink->reset(new stream_load::VOlapTableSinkV2(pool, row_desc, output_exprs, &status));
+            sink->reset(new vectorized::VOlapTableSinkV2(pool, row_desc, output_exprs, &status));
         } else {
-            sink->reset(new stream_load::VOlapTableSink(pool, row_desc, output_exprs, &status));
+            sink->reset(
+                    new vectorized::VOlapTableSink(pool, row_desc, output_exprs, false, &status));
         }
         RETURN_IF_ERROR(status);
         break;
@@ -315,8 +315,7 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
     case TDataSinkType::GROUP_COMMIT_OLAP_TABLE_SINK: {
         Status status;
         DCHECK(thrift_sink.__isset.olap_table_sink);
-        sink->reset(
-                new stream_load::GroupCommitVOlapTableSink(pool, row_desc, output_exprs, &status));
+        sink->reset(new vectorized::VOlapTableSink(pool, row_desc, output_exprs, true, &status));
         RETURN_IF_ERROR(status);
         break;
     }
