+++ be/src/vec/sink/vmysql_result_writer.h
@@ -51,7 +51,7 @@ public:
 
     bool can_sink() override;
 
-    Status close() override;
+    Status close(Status status) override;
 
     const ResultList& results() { return _results; }
