+++ be/src/vec/sink/writer/async_result_writer.h
@@ -48,9 +48,7 @@ class AsyncResultWriter : public ResultWriter {
 public:
     AsyncResultWriter(const VExprContextSPtrs& output_expr_ctxs);
 
-    Status close() override { return Status::OK(); }
-
-    void force_close();
+    void force_close(Status s);
 
     virtual bool in_transaction() { return false; }
 
@@ -92,7 +90,6 @@ private:
     std::deque<std::unique_ptr<Block>> _data_queue;
     Status _writer_status = Status::OK();
     bool _eos = false;
-    bool _force_close = false;
     bool _need_normal_close = true;
     bool _writer_thread_closed = false;
 };
